+++ pkg/analyzer/CHANGELOG.md
@@ -30,6 +30,7 @@
 * Deprecate `PropertyAccessorElement.isSynthetic`.
   Use `isOriginDeclaration`, `isOriginInterface`, `isOriginVariable` instead.
 * Deprecate `Element.isSynthetic`, use `isOriginXyz` properties instead.
+* Deprecate `LibraryElement.isSynthetic`, use `isOriginNotExistingFile` instead.
 
 ## 9.0.0
 * Remove deprecated field `Diagnostic.data`.